onboarding/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=5.0.0,<6.0.0
pulumi-command>=0.7.0,<0.7.5
pyyaml>=6.0.0,<6.5.0
pytest>=7.2.1,<7.2.2
requests>=2.28.2,<2.29.0